Add .travis.yml for continuous integration tests
Add a configuration file for [Travis CI][], which will run my tests
using the latest (development) version of the Go runtime as well as the
most recently stable version, currently Go 1.7.3.

Failure using Go tip will still result in the tests passing, since tip
is not a stable version. A failure on tip is still useful as it may
indicate that the next released version of Go is not compatible with the
code base:

https://docs.travis-ci.com/user/customizing-the-build/#Rows-that-are-Allowed-to-Fail

I've enabled `fast_finish` so that the build success/fails as soon as
the tests pass using Go 1.7.3:

https://docs.travis-ci.com/user/customizing-the-build/#Fast-Finishing

[Travis CI]: https://travis-ci.com