Clone Prometheus remote server code
Clone verbatim the example code for a Prometheus remote server:
https://github.com/prometheus/prometheus/blob/f002fe186abd04ecd9cd8961c24bc039f134ea13/documentation/examples/remote_storage/server.go

Plagarism disclaimer: For the purposes of my MSc project, let it be
noted that all of the code in this commit is copied from an external
source and none of it is my own work.

In my next commit, I'll adapt this code for use in Athens.