Add remote write endpoint to ingest samples
Add a 'remote write' endpoint that Prometheus can write to. For now,
samples are not stored and are simply printed to stdout.

The remove write server code is heavily based on the following example,
which was duplicated verbatim in my previous commit:
https://github.com/prometheus/prometheus/blob/f002fe186abd04ecd9cd8961c24bc039f134ea13/documentation/examples/remote_storage/server.go

The acceptance test for the remote write endpoint is also heavily based
on the code here:
https://github.com/prometheus/prometheus/blob/f002fe186abd04ecd9cd8961c24bc039f134ea13/storage/remote/client.go#L64-L115