Vendor remote.pb.go from Prometheus
Vendor the Protobuf structs used for samples we ingest from the remote
write API from the Prometheus project:
https://github.com/prometheus/prometheus/blob/254cb1ec29feb612fac87e26fcd8369ac4f91159/storage/remote/remote.pb.go

I'm vendoring the code outside `vendor/` because I'm about to start
using Prometheus libraries from the `dev-2.0` branch, and the `remote`
library I was using is absent from that branch (though I expect it'll be
added back later).

Plagarism disclaimer: For the purposes of my MSc project, let it be
noted that all of the code in this commit is copied from an external
source and none of it is my own work.