Persist ingested samples to local storage
Use the `github.com/prometheus/tsdb` library to persist ingested samples
to local disk.

See this blog post for an explanation of how the `tsdb` library works:
https://fabxc.org/blog/2017-04-10-writing-a-tsdb/

I've had to update the Prometheus packages I'm importing to use the
Prometheus `dev-2.0` branch so that the PromQL library is compatible
with the interfaces implemented by `github.com/prometheus/tsdb`.

For simplicity at this early stage in the project, I wiped the `vendor/`
directory and revendored all packages to pull in the Prometheus
`dev-2.0` libraries.