Add MkDocs stub for documentation and website
Use the MkDocs static site generator to add stub documentation.

Uses the Material theme for MkDocs because:

- it has a clean design
- the search functionality is very slick
- it has a responsive layout; works on phones/tablets
- it supports box-out notes and warnings
- it supports Markdown

https://github.com/squidfunk/mkdocs-material

To see a documentation in a development environment, use:

    make servedocs

Include building the documentation in strict mode as part of the `test`
make target to ensure that we have not broken the docs when making
changes. Add the generated `site` directory to `.gitignore`.

Add Docker as a dependency for Travis tests so that we can use Docker
to test the documentation builds correctly.

Pin the version of the `squidfunk/mkdocs-material` Docker image to
ensure the build is reproducible.