Decouple testing docs from app tests
I wanted to ensure that the documentation can always be rendered, so I
included it under the `test` Makefile target.

Because the `build` target depends on the `test` target,
the integration test environment added in 6eb62917bb was failing to
build due to the Docker dependency used to build the docs.

Rather than install Docker in the integration test environment, decouple
testing the documentation from the code itself, and ensure that Travis
tests both when a pull request is raised.