Pin the Docker Compose config version to 3.2
Specify the minor version for the Compose config file, in addition to
the major version, to avoid possible accidental breakage.

At the time of writing, the latest version is 3.3, but 3.2 is the latest
supported by my local version of `docker-compose`:

    docker-compose version 1.14.0, build c7bdf9e

I have no compelling reason to upgrade, and don't expect others to have
the very latest version, so I'll stick with version 3.2 for the Compose
config file for now.