Remove testdocs from the all Makefile target
Remove `testdocs` from the `all` Makefile target, otherwise I need to
support running Docker inside Docker when building the AthensDB Docker
image, which I'd rather avoid.

I could specify the Makefile targets to run in the AthensDB Dockerfile,
but I don't want to have to do that either. The `testdocs` target is run
in CI, and that's good enough for me.