Add integration test for Prometheus ingestion
Add an integration test that waits for Prometheus to send data to
AthensDB, then queries AthensDB for that data.

The integration tests run inside Docker Compose, which returns as soon
as the tests are complete with the exit code from the tests.

I've added two new Makefile targets:

- `clean` to remove the images built by Docker Compose
- `integration` to build and run the images specified in
`integration_tests/docker-compose.yml`

I'd normally avoid using a sleep in a test, but in this case it seems
justified - we want to wait long enough for Prometheus to scrape its own
metrics, i.e. a bit longer than the `scrape_interval` which happens at
the interval defined in `integration_tests/prometheus/prometheus.yml`.
We also have to wait for Prometheus to send the data to AthensDB.

I've used a build tag on the integration tests to prevent them from
being run by default.