Form a cluster using memberlist
Use Hashicorp's [memberlist][] library to form a cluster.

The AthensDB executable will read a comma-separated list of peers from
the `PEERS` environment variable and attempt to connect to those
peers.

So far the nodes will join form a cluster in as far as the nodes are
aware of the other nodes in the cluster, but no time-series data or
queries are exchanged between them yet.

I updated the integration tests to make the Prometheus instance send its
data to all 3 test nodes and have the integration tests query all 3
nodes. This doesn't add any benefit to the tests so far apart from
ensuring that I'm correctly spinning up 3 nodes and they are all able to
receive and query data.

I chose the memberlist library because it's already used by Hashicorp's
Serf and Consul products, which are widely deployed in production
environments, so it should be reasonably well tested and mature. It can
also communicate with potentially dead nodes by routing through known
good nodes when a network partition occurs.

The memberlist library uses a modified version of the [SWIM protocol][],
which uses TCP as well as a gossip-based UDP protocol to synchronise the
status of the cluster's nodes.

I considered using the more high-level (and more fully featured)
[Serf][] library, which builds on top of memberlist, however for now I
think that memberlist should be sufficient for my needs.

[memberlist]: https://github.com/hashicorp/memberlist
[SWIM protocol]: http://www.cs.cornell.edu/~asdas/research/dsn02-swim.pdf
[Serf]: https://github.com/hashicorp/serf