Forward writes to all other nodes in cluster
When receiving a write, forward it to all other nodes in the cluster.

Include a `X-AthensDB-Internal-Write-Version` HTTP header when writing
to other nodes to prevent them from trying to re-forward data to nodes
in an endless loop.

The code is a little rough but I plan to refactor it later and add more
tests.