Make vendored logrus library reflect vendor.json
I ran `govendor status` to check the integrity of vendored packages and
found that the vendored copy of the logrus library did not match the
version pinned in `vendor.json`.

govendor gave the following error:

    $ govendor status
    The following packages are missing or modified locally:
            github.com/Sirupsen/logrus
    Error: status failed for 1 package(s)

Additionally, remove the unused lowercase variant of logrus from
`vendor.json`. The URL for this library changed upstream to use
lowercase `sirupsen` instead of `Sirupsen`. The best way forward for now
is to continue to use the uppercase version as that's what's used by the
Prometheus libraries; mixed-case imports result in an import collision:

    can't load package: package github.com/sirupsen/logrus: case-insensitive import collision: "github.com/Sirupsen/logrus" and "github.com/sirupsen/logrus"

See this comment for more details:
https://github.com/sirupsen/logrus/issues/570#issuecomment-313933276