Fail tests if libraries missing or unused
Fail tests if libraries are missing from the vendor folder, if the
libraries in the vendor folder do not match the version in `vendor.json`
or if libraries are vendored but not actually used in the project.

Test for broken dependencies first in Travis so that tests will fail
fast as the `testdeps` target should return faster than `test` will.