Move main.go to 'cmd' directory
Move `main.go` to the `cmd` directory as is the common idiom in Go.

Since we don't want the acceptance tests to move into the `cmd`
directory, move them into their own package under
`internal/test/acceptance` and make the acceptance tests execute the
AthensDB binary rather than rely on running inside the `main` package.

A benefit of having the acceptance tests execute the binary is that the
acceptance tests can now also test command-line flags.

This also means the binary must now be built before the tests are run,
which is okay.