Dockerfile: Use scratch as base image
Instead of Alpine Linux, use the Docker `scratch` image as the base for
the final Docker image.

This image will now only contain the application executable
(`/athensdb`) and the `/etc/ssl/certs` directory copied from Debian,
used to verify certificate authorities for TLS connections to external
services.

Upgrade the `ca-certificates` package to ensure that it is as up-to-date
as possible.

This change reduces the Docker image size to 16.7MB, 78% of the previous
size (21MB).