Add ingestion routing
Route data samples received to the peers responsible for them. As a
first iteration, consider all peers to be responsible for every data
sample.

Use the data structure sampleNodeMap to group samples by node. I
considered just sending data verbatim to peer (forward the payload
as-is without any filtering) and let the receiving peer determine
which series to accept, but decided not to as:

a) the amount of additional network traffic could be significant

b) it would make it difficult to know which series could not be ingested
by the peers in the case of failure

I'll need to come back to this later to tidy the code and optimise it.