Cluster: Add hash ring
Add a hash ring using the `consistenthash` package that
deterministically partitions and replicates data samples across nodes in
the cluster.

Includes a test for a good distribution of data across all nodes in the
cluster, using 10,000 data samples spaced at 24-hour intervals. The
number 10,000 was chosen as a reasonable number of samples that would
not take too much time (or memory) to generate and test with.

The 24-hour interval is useful in testing the consistent hashing
algorithm and number of virtual nodes (determined by `hashringVnodes`),
rather than the partition key which hashes based on 24-hour intervals. I
should expand these tests later to better test the partition key.

A replication factor of 3 (specified by the constant
`replicationFactor`) means that all data in the cluster will be stored
in 3 copies, which in ideal conditions will be spread across the cluster
on at least 3 separate nodes, though that is not guaranteed by the
consistent hashing algorithm.

The replication factor is a trade-off between the financial cost of
storing additional copies versus the likelihood of losing data. The
choice of 3 replicas is based on industry norms. I have no plans to make
this constant configurable by the user in a bid to keep the
configuration simple.

Nodes leaving the cluster is not yet handled, I'll tackle that
separately.

See #27 for more details about the consistent hashing algorithm chosen
and see #12 for more details about the choice of partition key.

Closes #57.