Integration tests: Fail on application panic
The integration tests were failing due to a panic in the application but
the exit code was not being passed through to the Make target because
`--exit-code-from` was set.

Example where failing tests were ignored:
https://travis-ci.com/mattbostock/athensdb/jobs/88241024

In the above example, the application panicked before the integration
tests ran, and docker-compose returned a zero (`0`) exit code.

Remove the `--exit-code-from` option, which means that docker-compose
will return a non-zero exit code if any of the containers terminate with
non-zero exit code.

There's no need to know the precise exit code that the integration tests
returned, it's enough to know that the tests failed.