Rename GetNodesForSeries() to FilterBySeries()
Rename `GetNodesForSeries()` to `FilterBySeries()` and make it a method
on a new custom type, `Nodes`. The existing `Nodes()` function is
renamed to `GetNodes()` to avoid a naming conflict.

Making the function a method makes it easier to test the function by
passing in mock nodes.

`GetNodes()` is not an idiomatic name:
https://golang.org/doc/effective_go.html#Getters

...but it seemed cleaner to rename the function rather than the type
since the `Node` custom type also exists.