Cluster: Ensure replicas are evenly spread
Ensure that replicas of data are always spread evenly across the
cluster, so that each replica is always on a distinct node and no two
replicas reside on the same node.

The exception to this is on clusters which have fewer nodes than the
replication factor (as is the case in the acceptance tests), in which
case the replication factor is effectively capped at the number of nodes
in the cluster.

An even spread is achieved by 'forwarding' (figuratively, not literally)
samples to the next node in the hashring if a given node already
contains a replica of that sample. Continue to forward around the
ring until either all replicas have been placed, or, there are no more
nodes left to assign samples to. This algorithm is deterministic, so it
should not pose any problems for determining which nodes to query for
the samples on the read path.

Accordingly, refactor the cluster distribution tests to test several
cluster sizes to ensure to ensure test good coverage. The tests also now
call `FilterBySeries()`, which is the code being tested. Previously, the
tests were running a copy of the application code, which would mean
breakage could go undetected.

The tests now assert that replicas must be strictly evenly distributed.