Cluster: Reduce hashring vnodes
Reduce the number of vnodes used on the hashring from 1200 to 160.
Empirical testing using the tests in `cluster_test.go` suggests that the
distribution of samples and replicas across nodes, as determined using
the standard deviation of samples per node, is more even when using 160
vnodes.

I'll revisit this in future using more scientific evidence as a basis
for choosing the number of vnodes.