Integration tests: Build image just once
Set the `image` keyword to name the image built for `athensdb_1` and
have the other instances use that image. Previously, the image was built
separately (including test runs) for each instance.

Remove the `links` keyword from `athensdb_1` and use `depends_on` to
ensure that `athensdb_1` is built before the other instances run.

Also set the `--peers` flag on all instances to ensure that a cluster is
formed regardless of the order of instances starting.