Add benchmark utility and docker-compose env
Add a benchmarking environment that can be spun up ad-hoc in Docker
Compose.

Invoked by:

    make bench

The benchmark continuously sends the same 100,000 time-series to the
cluster via Nginx, incrementing the sample timestamp on each iteration.

Prometheus is included in the Docker Compose environment so that metrics
from AthensDB and the benchmark utility can be graphed.

I moved some of the functions from the `acceptance_test` to the
`testutil` package so that they can be reused for benchmarking.

The environment continues running even if one of the containers crashes,
which is useful for debugging issues on the running services and viewing
graphs in Prometheus after an issue occurred.

Nginx is configured to allow request bodies of 20MB to accommodate many
data samples in one request and buffering of request bodies to temporary
files is also disabled to help ensure good performance.