pprof: Include all profiling endpoints
The trace, cmdline, profile and symbol profilers were not previously
accessible as they are not included by `pprof.Index`. Add them
explicitly. These endpoints are extremely useful for debugging.

E.g.:

    go tool pprof --dot http://localhost:9080/debug/pprof/profile | dot -T png | open -f -a /Applications/Preview.app

...will open a diagram showing a CPU usage profile under macOS.

I tried the following to have the router reuse the handlers registered
automatically by the `pprof` package on the default servemux, however
it did not work as the route is matched exactly (not as a prefix).

```go
router.Get("/debug/pprof/", http.DefaultServeMux.ServeHTTP)
```