write: Protect tsdb appends with a mutex
From https://godoc.org/github.com/prometheus/tsdb#Appender:

> Operations on the Appender interface are not goroutine-safe.

It's not clear to me if this also means that multiple appenders are
unsafe for concurrent use, but with the latest version of the `tsdb`
library I'm using, a panic occurs if I don't protect the appender with a
mutex. It may be possible to relax this mutex later.