Preallocate slices on the hot path
Preallocate slices on the hot write path to save the CPU time the Go
runtime spends growing them.

In most cases it's known what the capacity of the slice should be. The
exception to that here is when unmarshalling data from a HTTP request,
in which case it's assumed that no more than 10,000 (`1e5`) time-series
will be received for the majority of the time.