bench: Run 3 goroutines to send samples
Run 3 goroutines to send data samples to nodes in the cluster, where 3
is the number of nodes in the benchmark cluster defined in
`docker-compose.yml`.

This is an attempt to try to saturate the capacity of all nodes.

The number `3` should be configurable to allow for benchmarking larger
clusters.

Adjust the size of the sampleChan buffer accordingly.