bench: Add metric for samples queued
Add a metric tracking the number of generated samples queued for
ingestion, which helps to determine if the samples are being generated
fast enough (currently, they are).

Add a `samplesToGenerate` constant to make it clear that `1e5` is the
number of samples being generated each time
`testutil.GenerateDataSamples()` is called.