bench: Remove Nginx and tweak load generation
Change how load is generated in `bench.go` and remove Nginx as the load
balancer used during the benchmark tests.

Remove Nginx because:
- it's not needed; the benchmark utility can round-robin between nodes
  in the cluster
- it's an extra dependency with associated configuration to maintain
- it's a potential bottleneck if not configured correctly
- it uses additional memory during benchmarking; I need as much memory
  as possible when benchmarking with my laptop
- it doesn't speak HTTP/2 by default (the Go http library does)

Instead, the bench utility now round robins between nodes in the
cluster. Removing Nginx yielded a (very approximate) 20% increase in
data sample throughput.

There's an argument that it's better to keep Nginx as it's likely to be
commonly used by end-users and it makes the benchmark more realistic,
but I don't see a lot of value in doing so. I want to benchmark AthensDB
rather than Nginx. The integration tests continue to include Nginx,
which is worth doing to confirm compatibility with AthensDB.

Load generation now uses multiple workers per node in the cluster,
taking the number of CPUs into account to try to generate as much load
as possible. I've removed the use of channels and instead count on the
fact that there's multiple workers to ensure that sample generation and
sample ingestion can occur in parallel.

By temporarily making the AthensDB `/receive` endpoint a no-op, I was
able to generate and send 500,000 samples per second during the
benchmarking tests on my laptop. This is faster than the cluster can
ingest currently, and the workload generation should scale to use more
CPU cores on larger machines.