write: Handle internal writes as fast as possible
Whenever a node forwards data to another node, it appends the
`HttpHeaderInternalWrite` HTTP header so that the receiving node knows
not to try to forward the data on (again).

This is a very common case: all data, assuming a replication factor
greater than 1, will traverse this path so make it as fast as
possible and return as soon as data is written.

Previously, internal writes occurred after determining which nodes
should store replicas of the data, which is expensive in terms of CPU.
Since internal writes are not replicated to other nodes, internal writes
should happen before then and return early.

The impact of this change is that nodes will blindly ingest any data
sent to them as an internal write; which I think for now is okay. I can
review this again when determining how to handle node failure in #33.