bench: Count number of unique series generated
Expose an approximation of the number of unique series generated by the
benchmark utility using the HyperLogLog++ algorithm.

A unique series is determined by the combination of all of its label
names and label values. Timestamp or sample values are not taken into
account.

This is a useful metric to track because the cardinality of series can
strongly correlate to memory usage and can be used when capacity
planning.

Use the HyperLogLog++ algorithm to keep the amount of memory used to
count unique series to a minimum. The number will not be precise, but
it's precise enough for this purpose:
http://static.googleusercontent.com/media/research.google.com/en/us/pubs/archive/40671.pdf