integration: Remove FIXME for data distribution
As of 17d6c5f4, this FIXME no longer applies. Replicas will always be
allocated to distinct node, so in the cluster of 3 nodes used by the
integration tests, each node will receive a replica so long as the
replication factor is set to 3 (which it is, as a constant).