Output logs as JSON
Output logs as JSON by default, so that log entries can be parsed in a
structured way.

I'd like to keep the colour human-readable that Logrus provides when a
TTY is in use, however for now the priority is to have JSON log output.
I may look at adding back more human-readable output in the near future.

I considered adding a configuration option for this, but I think the
JSON output is quite readable and I'd like to avoid adding configuration
options unless absolutely necessary.