write: Set partition key salt using HTTP header
Allow an optional HTTP header, `X-AthensDB-Partition-Key-Salt` to be set
in HTTP requests sending data to the cluster that adds a salt to the
partition key which can be used to better distribute data across the
cluster.

One use for the partition key salt is to support multiple tenants; by
using the tenant ID as the salt, data will be more evenly distributed if
multiple tenants try to write the same time-series.