Don't test with Go tip
The Travis CI tests run a little slowly which is not surprising as they
are integration tests.

To help reduce the time it takes for a CI job to complete, only test
with the latest stable Go version. New Go versions can be tested on a
branch as they are released.