Include metric in partition key
Include all label names and label values in the partition key. Since the
metric name is a label with the magic name `__name__`, the metric name
also forms part of the partition key.

This change will ensure that data is distributed more evenly across the
cluster. Specifically, it will avoid n nodes in the cluster receiving
all write traffic for a 24-hour period, where n is the replication
factor.

The partition key could now be represented as:

    <salt>:<bucket_end_time_as_YYYYMMDD>:<metric_name>:[<label_name>,<label_name>...]

This change makes it more difficult to reason about which nodes should
contain what data, but I think it's a necessary change to allow
horizontal scaling to all but the smallest of cluster sizes. With the
current schema, you might have a 19-node cluster but only 3 nodes will
ingest data for a given day. With this change, all 19 nodes will be
eligible to ingest data.

Another potential downside is that query aggregations across labels with
for the same metric name may now need to aggregate across multiple
nodes, however there may be a performance benefit from distributing the
query more widely. This could be optimised to some extent in future by
giving special treatment to labels that have special significant in
Prometheus such as `le` (for histograms), `job`, `quantile` (for
percentile summaries) and `instance`.

For more discussion on partition keys, see:
https://github.com/mattbostock/athensdb/issues/12

The partition schema used here is schema B in issue #12.

The time step when generated data samples in `distribution_test.go` is
reduced as it is no longer necessary to spread samples across days to
test how well distributed samples are across the cluster (the previous
partition schema dictated as much).