Rename project name from AthensDB to Timbala
The name AthensDB seems dull and unoriginal, plus I wanted to avoid
using 'DB' in the name, which was necessary to get a decent domain name.

Rename the project to 'Timbala', which is short and hopefully more
original. The word has no specific meaning that I'm aware of except that
it's similar to 'timbale', a type of drum, and I'm fond of drumming.

For now I've removed the old logo and favicon, I'll re-add new images
later.