Set limits on HTTP server
Set a 10MiB limit on the request body that the HTTP server will read, as
well as timeouts to prevent a misbehaving client from causing a denial
of service on the server. This change will also help to limit the amount
of memory that Timbala uses, since each HTTP request is read into memory
in its entirety.

If a client sends a request body larger than 10MiB and the
`Content-Length` request header indicates as much, a HTTP 413 "Request
Entity Too Large" response will be sent.

If no Content-Length header is set, the status code in the response is
dependent on how each handler handles an error when reading from the
response body.

An improvement would be to always ensure a HTTP 413 is returned
whether a Content-Length header is set or not, but the current
implementation of `http.MaxBytesReader()` would require matching a
specific error string, which is somewhat brittle. I decided it's not
worth the additional complexity right now.

Adds a test to check if a HTTP 413 is returned for oversize requests to
the read and write APIs.