Be consistent about errors reading request body
Make the write API consistent with the read API and return a "Bad
Request" HTTP response status when an error is encountered while reading
from the request body.

I'm not yet sure if this should be a "Bad Request" or an "Internal
Server Error" but since 5XX status codes are used for monitoring, err on
the side of less noise.