read: Fan out remote reads across the cluster
When requests are sent to the remote read API on the `/read` HTTP
endpoint, forward the query to all nodes in the cluster (fan out) and
merge the results before returning the response.

Previously, nodes would only respond with data that they had in their
local storage.

Adds a new `fanout` package which implements the Prometheus
`storage.Storage` interface and can eventually be used for writes as
well as reads.