integration: Only write to 2 nodes during fanout
When testing the fanout of read queries across the cluster (i.e. if
queries are sent to multiple nodes, not just the local node), only write
directly to 2 nodes out of the 3 nodes in the integration test cluster.

Doing so helps ensure that the tests are not passing just because the
writes are accidentally being repeated (fanned out) to all nodes in the
cluster (in which case all nodes would have the data being queried).

Additionally, send queries to the third node in the cluster which is not
written to during these tests.

Tidies up the test assertions by memoising expected versus 'got' results
to avoid duplicate code.